backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector==0.3.6

# AI/ML
sentence-transformers==3.3.1
torch==2.5.1
kiwipiepy==0.21.0
scikit-learn==1.6.0
numpy==2.2.1

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.7.0
email-validator==2.2.0

# HTTP Client
httpx==0.28.1

# Redis (for caching)
redis==5.2.1
hiredis==3.0.0

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
ruff==0.8.4
mypy==1.14.0
